initial wip shared report templates
